requirements.txt
ahk
numpy
pandas
pillow
scikit-learn
scikit-image